created small nav bar component, removed console logs
