+++ src/components/App/App.css
@@ -723,6 +723,7 @@ html {
   background: linear-gradient(0deg, rgba(37,37,37,1) 0%, #333333 100%);
   z-index: 100;
   border: 1px solid #333333;
+  border-bottom: none;
 };
